requirements.txt
aiogram==3.15.0
python-dotenv==1.0.1
openpyxl==3.1.5
supabase==2.11.0